[script] add cast String <->[Byte]

+++ src/xci/script/Machine.cpp
@@ -37,7 +37,7 @@ void Machine::call(const Function& function, const Machine::InvokeCallback& cb)
 
 void Machine::run(const InvokeCallback& cb)
 {
-    // Avoid double-recursion - move these pointers instead (we already have stack)
+    // Avoid double-recursion - move these pointers instead (we already have a stack)
     const Function* function = &m_stack.frame().function;
     auto it = function->code().begin() + (ptrdiff_t) m_stack.frame().instruction;
     auto base = m_stack.frame().base;
@@ -298,7 +298,7 @@ void Machine::run(const InvokeCallback& cb)
                 auto arg = m_stack.pull_typed(ti_list(TypeInfo(elem_ti)));
                 auto len = arg.get<ListV>().length();
                 arg.decref();
-                m_stack.push(value::UInt32(len));
+                m_stack.push(value::UInt32(uint32_t(len)));
                 break;
             }
